pyproject.toml
[tool.poetry]
name = "image-search"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.dependencies]
python = ">3.9,<3.13"
Flask = "^3.0.3"
chromadb = "^0.5.3"
Werkzeug = "^3.0.3"
numpy = ">=1.22.5,<2.0.0"
pillow = "^10.3.0"
osxphotos = "^0.68.2"
mypy = "^1.10.0"
open-clip-torch = "^2.24.0"
isort = "^5.13.2"
black = "^24.4.2"
